Feat/subnet per nodepool (#354)
* Move template functions to `templateUtils`

* Add test

* Add new function to template parser

* Subnets for azure tpl

* Add subnets for AWS tpl

* Add subnets for OCI

* Add subnets for GCP tpl

* Auto commit - update kustomization.yaml

* Auto commit - update kustomization.yaml

* Auto commit - update kustomization.yaml
