+++ internal/templateUtils/template_test.go
@@ -1,0 +1,8 @@
+package templateUtils
+
+import "testing"
+
+func TestGetCIDR(t *testing.T) {
+	cidr := GetCIDR("10.0.0.0/16", 2, 1)
+	t.Log(cidr)
+}
