+++ internal/templateUtils/templates.go
@@ -8,7 +8,6 @@ import (
 
 	"text/template"
 
-	"github.com/Berops/claudie/internal/utils"
 	"github.com/rs/zerolog/log"
 )
 
@@ -72,11 +71,12 @@ func (t Templates) GenerateToString(tpl *template.Template, d interface{}) (stri
 func (tl TemplateLoader) LoadTemplate(tplFile string) (*template.Template, error) {
 	tpl := template.New(tplFile).
 		Funcs(template.FuncMap{
-			"isMissing":                     utils.IsMissing[int],
-			"targetPorts":                   utils.ExtractTargetPorts,
-			"protocolToOCIProtocolNumber":   utils.ProtocolNameToOCIProtocolNumber,
-			"protocolToAzureProtocolString": utils.ProtocolNameToAzureProtocolString,
-			"assignPriority":                utils.AssignPriority,
+			"isMissing":                     IsMissing[int],
+			"targetPorts":                   ExtractTargetPorts,
+			"protocolToOCIProtocolNumber":   ProtocolNameToOCIProtocolNumber,
+			"protocolToAzureProtocolString": ProtocolNameToAzureProtocolString,
+			"assignPriority":                AssignPriority,
+			"getCIDR":                       GetCIDR,
 		})
 
 	tpl, err := tpl.ParseFiles(filepath.Join(baseDirectory, tl.Directory, tplFile))
