+++ manifests/claudie/kustomization.yaml
@@ -21,18 +21,18 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: claudieio/ansibler
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/builder
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/context-box
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/frontend
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/kube-eleven
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/kuber
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/scheduler
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
 - name: claudieio/terraformer
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
