+++ manifests/testing-framework/kustomization.yaml
@@ -22,4 +22,4 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 images:
 - name: claudieio/testing-framework
-  newTag: 49eb227-1041
+  newTag: 561b5f8-1045
